[hw1] debug

+++ src/hw1/resize_image.c
@@ -12,9 +12,9 @@ image nn_resize(image im, int w, int h)
 {
     // TODO Fill in (also fix that first line)
     image result = make_image(w, h, im.c);
-    float width_ratio = im.w / (float)w;
+    float width_ratio = (float) im.w / w;
     float width_off = 0.5 * (width_ratio - 1);
-    float height_ratio = im.h / (float)h;
+    float height_ratio = (float) im.h / h;
     float height_off = 0.5 * (height_ratio - 1);
     for (int c = 0; c < 3; c++) {
         for (int y = 0; y < result.h; y++) {
@@ -31,13 +31,13 @@ image nn_resize(image im, int w, int h)
 
 float bilinear_interpolate(image im, float x, float y, int c)
 {
-    int x_min = (int) floorf(x);
-    int x_max = x_min + 1;
-    int y_min = (int) floorf(y);
-    int y_max = y_min + 1;
+    float x_min = floor(x);
+    float x_max = x_min + 1;
+    float y_min = floor(y);
+    float y_max = y_min + 1;
 
-    float q1 = (y - y_min) * get_pixel(im, x_min, y_min, c) + (y_max - y) * get_pixel(im, x_min, y_max, c);
-    float q2 = (y - y_min) * get_pixel(im, x_max, y_min, c) + (y_max - y) * get_pixel(im, x_max, y_max, c);
+    double q1 = (y - y_min) * get_pixel(im, x_min, y_min, c) + (y_max - y) * get_pixel(im, x_min, y_max, c);
+    double q2 = (y - y_min) * get_pixel(im, x_max, y_min, c) + (y_max - y) * get_pixel(im, x_max, y_max, c);
 
     return (x - x_min) * q1 + (x_max - x) * q2;
 }
@@ -46,9 +46,9 @@ image bilinear_resize(image im, int w, int h)
 {
     // TODO
     image result = make_image(w, h, im.c);
-    float width_ratio = im.w / (float)w;
+    float width_ratio = (float) im.w / w;
     float width_off = 0.5 * (width_ratio - 1);
-    float height_ratio = im.h / (float)h;
+    float height_ratio = (float) im.h / h;
     float height_off = 0.5 * (height_ratio - 1);
     for (int c = 0; c < 3; c++) {
         for (int y = 0; y < result.h; y++) {
